pyproject.toml
[project]
name = "www-pasic-robinson-org"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "ruff>=0.14.10",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.lint.per-file-ignores]
"build.py" = ["E501", "SIM115"] # build.py